POPRAWIONE: unikalne usernamy, unikalne nazwy dań, dodano roota, obsługa usunięcia usera który jest zalogowany (chyba), wszystkie Get-y poprawnie zwracają tylko nieusuniętych userów. BEZ ZMIAN: obsługa błędów w formularzu do tworzenia/edycji dań (cena, czas trwania)

+++ restauracja-wpf/ModifyUserWindow.xaml.cs
@@ -1,4 +1,5 @@
 ﻿using restauracja_wpf.Data;
+using restauracja_wpf.Interfaces;
 using restauracja_wpf.Models;
 using restauracja_wpf.Services;
 using System;
@@ -22,7 +23,7 @@ namespace restauracja_wpf
     /// </summary>
     public partial class ModifyUserWindow : Window
     {
-        private readonly UserDataService userService = new (new GenericDataService<User>(new RestaurantContextFactory()));
+        private readonly UserDataService userService = new(new GenericDataService<User>(new RestaurantContextFactory()));
         private readonly User user;
 
         public ModifyUserWindow(User user)
@@ -36,27 +37,37 @@ namespace restauracja_wpf
             txtLogin.Text = user.Login;
             FillUpComboboxAsync();
             cmbRole.SelectedValue = user.RoleId;
-            cmbRestaurant.SelectedValue = user.RestaurantId;
+            //cmbRestaurant.SelectedValue = user.RestaurantId;
             ckbEnabled.IsChecked = user.Status;
         }
 
         private async void FillUpComboboxAsync()
         {
-            GenericDataService<Role> roleService = new GenericDataService<Role>(new RestaurantContextFactory());
-            cmbRole.ItemsSource = await roleService.GetAll();
+            RoleDataService roleService = new(new GenericDataService<Role>(new RestaurantContextFactory()));
+            if (SessionManager.CurrentUser.Role.Name == "root")
+                cmbRole.ItemsSource = await roleService.GetAllRoles(true);
+            else
+                cmbRole.ItemsSource = await roleService.GetAllRoles();
+
             cmbRole.DisplayMemberPath = "Name";
             cmbRole.SelectedValuePath = "Id";
             cmbRole.SelectedIndex = 0;
 
-            GenericDataService<Restaurant> restaurantService = new GenericDataService<Restaurant>(new RestaurantContextFactory());
-            cmbRestaurant.ItemsSource = await restaurantService.GetAll();
-            cmbRestaurant.DisplayMemberPath = "Name";
-            cmbRestaurant.SelectedValuePath = "Id";
-            cmbRestaurant.SelectedIndex = 0;
+            //GenericDataService<Restaurant> restaurantService = new GenericDataService<Restaurant>(new RestaurantContextFactory());
+            //cmbRestaurant.ItemsSource = await restaurantService.GetAll();
+            //cmbRestaurant.DisplayMemberPath = "Name";
+            //cmbRestaurant.SelectedValuePath = "Id";
+            //cmbRestaurant.SelectedIndex = 0;
         }
 
         private async void btnConfirmChanges_Click(object sender, RoutedEventArgs e)
         {
+            if (user.Role.Name == "root" || user.Role.Name == "admin")
+            {
+                MessageBox.Show("Only root can manage this account.", "Error", MessageBoxButton.OK, MessageBoxImage.Error);
+                return;
+            }
+
             if (MessageBox.Equals(MessageBox.Show("Are you sure you want to update user details?", "Confirm Update", MessageBoxButton.YesNo, MessageBoxImage.Question), MessageBoxResult.No))
             {
                 return; // User cancelled the update
@@ -71,15 +82,23 @@ namespace restauracja_wpf
                 MessageBox.Show("Passwords do not match.", "Validation Error", MessageBoxButton.OK, MessageBoxImage.Error);
                 return; // throw new Exception("Passwords do not match.");
             }
-            else if (cmbRestaurant.SelectedValue == null)
-            {
-                MessageBox.Show("Please select a restaurant.", "Validation Error", MessageBoxButton.OK, MessageBoxImage.Error);
-                return; // throw new Exception("Please select a restaurant.");
-            }
+            //else if (cmbRestaurant.SelectedValue == null)
+            //{
+            //    MessageBox.Show("Please select a restaurant.", "Validation Error", MessageBoxButton.OK, MessageBoxImage.Error);
+            //    return; // throw new Exception("Please select a restaurant.");
+            //}
             else if (cmbRole.SelectedValue == null)
             {
                 MessageBox.Show("Please select a role.", "Validation Error", MessageBoxButton.OK, MessageBoxImage.Error);
                 return; // throw new Exception("Please select a role.");
+            }
+
+            var usernameTaken = await userService.IsUsernameTaken(txtLogin.Text);
+
+            if (usernameTaken)
+            {
+                MessageBox.Show("This username is already taken. Please choose a different one.", "Validation Error", MessageBoxButton.OK, MessageBoxImage.Error);
+                return; // throw new Exception("This login is already taken.");
             }
 
             try
@@ -91,7 +110,7 @@ namespace restauracja_wpf
                     Login = txtLogin.Text,
                     PasswordHash = txtChangePassword.Password,
                     RoleId = Convert.ToInt32(cmbRole.SelectedValue),
-                    RestaurantId = Convert.ToInt32(cmbRestaurant.SelectedValue),
+                    RestaurantId = null!,
                     Status = ckbEnabled.IsChecked == true ? true : false
                 });
                 MessageBox.Show("User details updated successfully.", "Success", MessageBoxButton.OK, MessageBoxImage.Information);
@@ -105,6 +124,17 @@ namespace restauracja_wpf
 
         private async void btnDeleteAccount_Click(object sender, RoutedEventArgs e)
         {
+            if (user.Role.Name == "root" || user.Role.Name == "Admin")
+            {
+                MessageBox.Show("Only root can manage this account.", "Error", MessageBoxButton.OK, MessageBoxImage.Error);
+                return;
+            }
+            else if (SessionManager.CurrentUser.Role.Name == "root")
+            {
+                MessageBox.Show($"You can't delete the root account.", "Error", MessageBoxButton.OK, MessageBoxImage.Error);
+                return;
+            }
+
             if (MessageBox.Equals(MessageBox.Show($"DELETE user account?", "Confirm", MessageBoxButton.OKCancel, MessageBoxImage.Question), MessageBoxResult.OK))
             {
                 await userService.Delete(user.Id);
